[refactor/#229] 컨버터 로직 dto로 이전
